requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
qrcode==7.4.2
Pillow==10.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0